requirement.txt
django
django-cleanup
pillow